src/backend/Documentos/requirements.txt
fpdf2
psycopg2-binary
boto3
fpdf